chore: moves slice to store folder
