+++ src/components/Navigation.js
@@ -28,7 +28,7 @@ import {
   UNLEASH_TOKEN_BALANCES_FEATURE_FLAG,
   REACT_APP_NETWORK,
 } from '../constants';
-import { toggleTheme } from '../reducers';
+import { toggleTheme } from '../store/defaultSlice';
 import NewHathorAlert from './NewHathorAlert';
 
 function Navigation() {
