+++ src/components/ThemeSwitch.js
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { toggleTheme } from '../reducers';
+import { toggleTheme } from '../store/defaultSlice';
 
 export const ThemeSwitch = () => {
   const theme = useSelector(state => state.theme);
