+++ src/store/index.js
@@ -6,7 +6,7 @@
  */
 
 import { configureStore } from '@reduxjs/toolkit';
-import rootReducer from '../reducers/index';
+import rootReducer from './defaultSlice';
 
 const store = configureStore({ reducer: rootReducer });
